Share how the blocks of a feature file are found
The loop walking a feature file for docstrings existed twice: as
`collect_blocks()` in the analysis and inlined in the extractor of the code
style check, which is where the two had already come to disagree about what
counts as a PHP block. The check needs the narrower rule, since it writes a
block back and reformatting an expectation about the contents of a file
would make it stop matching, while the analysis wants every block it can
read.

Rather than a flag deciding that inside the shared loop, a block now comes
back saying which rule recognised it and whether it brings its own opening
tag, and each tool takes it from there. That also drops the separate pass
the check made over a block to answer the second question.

The walk over the feature files of a directory is shared along with it. The
code style check gains the stable order the analysis already relied on.
